+++ src/main/java/roomescape/reservation/domain/repository/ReservationRepository.java
@@ -1,9 +1,11 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.repository;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import roomescape.member.domain.Member;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationSlot;
 import roomescape.theme.domain.Theme;
 
 public interface ReservationRepository {
