+++ src/main/java/roomescape/reservation/domain/repository/ReservationTimeRepository.java
@@ -1,8 +1,9 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.repository;
 
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
+import roomescape.reservation.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
