+++ src/main/java/roomescape/reservation/domain/repository/WaitingRepository.java
@@ -1,8 +1,11 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
 import roomescape.member.domain.Member;
+import roomescape.reservation.domain.ReservationSlot;
+import roomescape.reservation.domain.Waiting;
+import roomescape.reservation.domain.WaitingWithRank;
 
 public interface WaitingRepository {
