+++ src/main/java/roomescape/reservation/infrastructure/ReservationRepositoryImpl.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReservationSlot;
+import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 
 @Repository
