+++ src/main/java/roomescape/reservation/infrastructure/WaitingRepositoryImpl.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.ReservationSlot;
 import roomescape.reservation.domain.Waiting;
-import roomescape.reservation.domain.WaitingRepository;
 import roomescape.reservation.domain.WaitingWithRank;
+import roomescape.reservation.domain.repository.WaitingRepository;
 
 @Repository
 @RequiredArgsConstructor
