+++ src/test/java/roomescape/reservation/domain/ReservationRepositoryTest.java
@@ -20,6 +20,8 @@ import org.springframework.context.annotation.Import;
 import roomescape.fixture.config.TestConfig;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRepository;
+import roomescape.reservation.domain.repository.ReservationRepository;
+import roomescape.reservation.domain.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeRepository;
