+++ src/test/java/roomescape/theme/domain/ThemeRepositoryTest.java
@@ -18,10 +18,10 @@ import roomescape.fixture.domain.ThemeFixture;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRepository;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReservationSlot;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.domain.ReservationTimeRepository;
+import roomescape.reservation.domain.repository.ReservationRepository;
+import roomescape.reservation.domain.repository.ReservationTimeRepository;
 
 @DataJpaTest
 @Import(TestConfig.class)
